refactor smart contract interface
